Extract serialization template functions into a serialization Jinja2 extension
